+++ docs/make.jl
@@ -79,6 +79,7 @@ makedocs(
             "howtos/6_sinking_particles.md",
             "howtos/7_etopo.md",
             "howtos/8_nonlinearsolve.md",
+            "howtos/9_timestepping.md",
             "howtos/analytical_derivatives.md",
         ],
         "Explanation" => [